Move internal headers to a subdirectory
The only rocsolver header intended for direct inclusion by users is
<rocsolver/rocsolver.h>. The other rocsolver headers were not
explicitly marked as internal, but that is what this change is fixing.
All headers within internal/ are implementation details and may change
without warning.
