+++ library/CMakeLists.txt
@@ -6,15 +6,15 @@
 set(rocsolver_SOVERSION 0.1)
 
 # Create version header from templated .in file using CMake info
-configure_file(include/rocsolver/rocsolver-version.h.in "${PROJECT_BINARY_DIR}/include/rocsolver/rocsolver-version.h")
+configure_file(include/rocsolver/internal/rocsolver-version.h.in "${PROJECT_BINARY_DIR}/include/rocsolver/internal/rocsolver-version.h")
 
 set(rocsolver_headers_public
   include/rocsolver/rocsolver.h
-  include/rocsolver/rocsolver-extra-types.h
-  include/rocsolver/rocsolver-aliases.h
-  include/rocsolver/rocsolver-functions.h
-  ${PROJECT_BINARY_DIR}/include/rocsolver/rocsolver-version.h
-  ${PROJECT_BINARY_DIR}/include/rocsolver/rocsolver-export.h
+  include/rocsolver/internal/rocsolver-extra-types.h
+  include/rocsolver/internal/rocsolver-aliases.h
+  include/rocsolver/internal/rocsolver-functions.h
+  ${PROJECT_BINARY_DIR}/include/rocsolver/internal/rocsolver-version.h
+  ${PROJECT_BINARY_DIR}/include/rocsolver/internal/rocsolver-export.h
 )
 
 # For IDE integration
