+++ library/include/rocsolver/internal/rocsolver-extra-types.h
