+++ library/include/rocsolver/internal/rocsolver-version.h.in
